Feat : 전략 패턴 적용 (#3)
