+++ src/main/java/com/utopia/utopia_be/auth/service/AuthService.java
@@ -4,7 +4,6 @@ import static com.utopia.utopia_be.auth.exception.errorcode.AuthErrorCode.*;
 import static com.utopia.utopia_be.user.exception.errorcode.UserErrorCode.*;
 
 import java.util.Map;
-import java.util.function.Function;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -14,11 +13,10 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
 import com.utopia.utopia_be.auth.dto.request.BasicLoginRequest;
-import com.utopia.utopia_be.auth.dto.response.KakaoUserInfoResponse;
 import com.utopia.utopia_be.auth.dto.response.LoginResponse;
 import com.utopia.utopia_be.auth.exception.AuthException;
+import com.utopia.utopia_be.auth.service.strategy.SocialLoginStrategy;
 import com.utopia.utopia_be.auth.util.JwtTokenProvider;
-import com.utopia.utopia_be.auth.util.kakao.KakaoApiClient;
 import com.utopia.utopia_be.user.domain.User;
 import com.utopia.utopia_be.user.domain.type.LoginType;
 import com.utopia.utopia_be.user.exception.UserException;
@@ -32,56 +30,19 @@ public class AuthService {
 
   private final UserRepository userRepository;
   private final JwtTokenProvider jwtTokenProvider;
-  private final KakaoApiClient kakaoApiClient;
-
   private final PasswordEncoder passwordEncoder;
 
-  private final Map<LoginType, Function<String, LoginResponse>> loginStrategyMap =
-      Map.of(
-          LoginType.KAKAO, this::kakaoLogin,
-          LoginType.NAVER, this::naverLogin,
-          LoginType.GOOGLE, this::googleLogin);
+  private final Map<String, SocialLoginStrategy> loginStrategyMap;
 
   public LoginResponse socialLogin(LoginType loginType, String code) {
-    Function<String, LoginResponse> strategy = loginStrategyMap.get(loginType);
-    if (strategy == null) {
+    SocialLoginStrategy loginStrategy =
+        loginStrategyMap.get(loginType.name().toLowerCase() + "LoginStrategy");
+
+    if (loginStrategy == null) {
       throw new AuthException(LOGIN_TYPE_NOT_SUPPORTED);
     }
-    return strategy.apply(code);
-  }
 
-  private LoginResponse kakaoLogin(String code) {
-    // Authorization Code를 Access Token으로 교환
-    String accessToken = kakaoApiClient.getAccessToken(code);
-
-    // Access Token을 이용해 사용자 정보를 가져오고 없으면 회원가입
-    KakaoUserInfoResponse userInfo = kakaoApiClient.getUserInfo(accessToken);
-    User user = findOrCreateUser(userInfo.kakao_account().email(), LoginType.KAKAO);
-
-    // JWT 토큰 생성
-    String token = jwtTokenProvider.createToken(user.getId().toString());
-
-    return LoginResponse.from(user.getNickname(), user.getRole(), token);
-  }
-
-  private LoginResponse naverLogin(String code) {
-    return null;
-  }
-
-  private LoginResponse googleLogin(String code) {
-    return null;
-  }
-
-  public User findOrCreateUser(String email, LoginType loginType) {
-    return userRepository
-        .findByEmailAndLoginType(email, loginType)
-        .orElseGet(
-            () ->
-                userRepository.save(
-                    User.socialLoginBuilder()
-                        .email(email)
-                        .loginType(loginType)
-                        .buildSocialLogin()));
+    return loginStrategy.login(code);
   }
 
   public void signup(BasicLoginRequest request) {
