+++ src/main/java/com/utopia/utopia_be/auth/service/strategy/GoogleLoginStrategy.java
@@ -1,0 +1,14 @@
+package com.utopia.utopia_be.auth.service.strategy;
+
+import org.springframework.stereotype.Service;
+
+import com.utopia.utopia_be.auth.dto.response.LoginResponse;
+
+@Service
+public class GoogleLoginStrategy implements SocialLoginStrategy {
+
+  @Override
+  public LoginResponse login(String code) {
+    return null;
+  }
+}
