+++ src/main/java/com/utopia/utopia_be/auth/service/strategy/KakaoLoginStrategy.java
@@ -1,0 +1,39 @@
+package com.utopia.utopia_be.auth.service.strategy;
+
+import org.springframework.stereotype.Service;
+
+import lombok.RequiredArgsConstructor;
+
+import com.utopia.utopia_be.auth.dto.response.KakaoUserInfoResponse;
+import com.utopia.utopia_be.auth.dto.response.LoginResponse;
+import com.utopia.utopia_be.auth.util.JwtTokenProvider;
+import com.utopia.utopia_be.auth.util.kakao.KakaoApiClient;
+import com.utopia.utopia_be.user.domain.User;
+import com.utopia.utopia_be.user.domain.type.LoginType;
+import com.utopia.utopia_be.user.service.UserService;
+
+@Service
+@RequiredArgsConstructor
+public class KakaoLoginStrategy implements SocialLoginStrategy {
+
+  private final KakaoApiClient kakaoApiClient;
+  private final UserService userService;
+  private final JwtTokenProvider jwtTokenProvider;
+
+  @Override
+  public LoginResponse login(String code) {
+    // 1. Authorization Code를 Access Token으로 교환
+    String accessToken = kakaoApiClient.getAccessToken(code);
+
+    // 2. Access Token을 이용해 사용자 정보를 가져오고 없으면 회원가입
+    KakaoUserInfoResponse userInfo = kakaoApiClient.getUserInfo(accessToken);
+
+    // 3. 사용자 정보로 회원가입 또는 로그인
+    User user = userService.findOrCreateUser(userInfo.kakao_account().email(), LoginType.KAKAO);
+
+    // 4. JWT 토큰 생성
+    String token = jwtTokenProvider.createToken(user.getId().toString());
+
+    return LoginResponse.from(user.getNickname(), user.getRole(), token);
+  }
+}
